+++ addons/web/static/src/js/action_manager/action_plugins/report_action_plugin.js
@@ -6,6 +6,7 @@ odoo.define('web.ReportActionPlugin', function (require) {
      * 'ir.actions.report' to the ActionManager.
      */
 
+    const AbstractActionPlugin = require('web.AbstractActionPlugin');
     const ActionManager = require('web.ActionManager');
     const core = require('web.core');
     const session = require('web.session');
@@ -27,7 +28,7 @@ odoo.define('web.ReportActionPlugin', function (require) {
                      'the reports.'),
     };
 
-    class ReportActionPlugin extends ActionManager.AbstractPlugin {
+    class ReportActionPlugin extends AbstractActionPlugin {
 
         //----------------------------------------------------------------------
         // Public
