+++ addons/web/static/src/js/action_manager/action_plugins/window_action_plugin.js
@@ -1,16 +1,16 @@
-odoo.define('web.ActWindowActionManager', function (require) {
+odoo.define('web.WindowActionPlugin', function (require) {
     "use strict";
 
     /**
      * The purpose of this file is to add the support of Odoo actions of type
      * 'ir.actions.act_window' to the ActionManager.
      */
-
+    const AbstractActionPlugin = require('web.AbstractActionPlugin');
     const ActionManager = require('web.ActionManager');
     const { action_registry } = require('web.core');
     const viewRegistry = require('web.view_registry');
 
-    class WindowActionPlugin extends ActionManager.AbstractPlugin {
+    class WindowActionPlugin extends AbstractActionPlugin {
         constructor() {
             super(...arguments);
             this.env.bus.on('switch-view', this, this._onSwitchView);
